+++ maven-ems/src/main/java/com/ideas2it/ems/employee/service/EmployeeServiceImpl.java
@@ -1,13 +1,11 @@
 package com.ideas2it.ems.employee.service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.time.LocalDate;
+import java.util.Set;
 
 import com.ideas2it.ems.model.BankDetail;
-import com.ideas2it.ems.bankDetail.service.BankDetailService;
-import com.ideas2it.ems.bankDetail.service.BankDetailServiceImpl;
-import com.ideas2it.ems.certificate.service.CertificateService;
-import com.ideas2it.ems.certificate.service.CertificateServiceImpl;
 import com.ideas2it.ems.model.Certificate;
 import com.ideas2it.ems.model.Department;
 import com.ideas2it.ems.department.service.DepartmentService;
@@ -15,8 +13,10 @@ import com.ideas2it.ems.department.service.DepartmentServiceImpl;
 import com.ideas2it.ems.employee.dao.EmployeeDao;
 import com.ideas2it.ems.employee.dao.EmployeeDaoImpl;
 import com.ideas2it.ems.model.Employee;
-import com.ideas2it.ems.employee.service.EmployeeService;
 import com.ideas2it.ems.exception.EmployeeException;
+
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
 
 /**
  * Service class for managing employee-related operations.
@@ -27,7 +27,8 @@ import com.ideas2it.ems.exception.EmployeeException;
 public class EmployeeServiceImpl implements EmployeeService {
     EmployeeDao employeeDao = new EmployeeDaoImpl();
     DepartmentService departmentService = new DepartmentServiceImpl();
-    BankDetailService bankDetailService = new BankDetailServiceImpl();
+
+    private static final Logger logger = LogManager.getLogger();
 
     @Override
     public void addEmployee(String employeeName, LocalDate employeeDOB, long contactNumber, String mailId, int experience, 
@@ -51,10 +52,7 @@ public class EmployeeServiceImpl implements EmployeeService {
     @Override
     public boolean isPresentEmployee(int employeeId) throws EmployeeException {
         Employee employee = employeeDao.getEmployeeById(employeeId);
-        if (employee == null) {
-            return false;
-        } 
-        return true;
+        return employee != null;
     }
 
     @Override
@@ -63,13 +61,27 @@ public class EmployeeServiceImpl implements EmployeeService {
     }
 
     @Override
-    public void deleteEmployee(int employeeId) throws EmployeeException {
-        employeeDao.deleteEmployee(employeeId);
+    public boolean deleteEmployee(int employeeId) throws EmployeeException {
+        Employee employee = employeeDao.getEmployeeById(employeeId);
+        if(null == employee) {
+            logger.warn("Given employee id does not exit {}" , employeeId);
+            return false;
+        } else {
+            employeeDao.deleteEmployee(employeeId);
+            return true;
+        }
     }
 
     @Override
     public List<Certificate> getCertificatesByEmployeeId(int employeeId) throws EmployeeException {
-        return employeeDao.getCertificatesByEmployeeId(employeeId);
+        Employee employee = employeeDao.getEmployeeById(employeeId);
+        List<Certificate> certificates = null;
+        if(null == employee) {
+            logger.warn("Given employee id does not exit {}" , employeeId);
+        } else {
+            certificates = new ArrayList<>(employee.getCertificates());
+        }
+        return certificates;
     }
  
     @Override
